Fix issues with parts of env line not being obfuscated
When dealing with such line to obfuscate:

    FOO=a b c BAR=d

env vars FOO and BAR are technically keeping values a and d. Following
such interpretetion we were obfuscating only those values. This is wrong
when it comes to security - it's easy to forget to put quotes for the
first variable and in such case values b and c will be leaked.

This commit fixes such situations and strengthens test suite for
obfuscation.